src/engine/mtrl.cpp: use uintptr_t case for identifier field access

+++ src/engine/mtrl.cpp
@@ -311,7 +311,7 @@ PMTRL MTRL::PmtrlFromBmtl(PBMTL pbmtl)
 {
     AssertVarMem(pbmtl);
 
-    PMTRL pmtrl = (PMTRL) * (int32_t *)pbmtl->identifier;
+    PMTRL pmtrl = (PMTRL) * (uintptr_t *)pbmtl->identifier;
     AssertPo(pmtrl, 0);
     return pmtrl;
 }
